5:51 Note: when I add px-6;  (Rems), Inspect->layout console. 4:00 - 5:33
... tailwind is going to add on the x-axis padding left and right of 1.5 rems. rem: root em; the root font size.
And if that's 16 as it usually is. 1.5rems will be 16*1.5 = 24...
What's cool is if it's based on the room em, If change the font size here it will cascade down and update all of these references.
So let's do sth crazy like 50px on the html
inspect...
element {
    font-size: 50px;
}
Now it changes the padding on the html... to 75px.
50 * 1.5 = 75.

Often do this to control font sizes for say mobile.
So I could say maybe text is small; text-sm or base; text-base on mobile but when u get to a large size it will be extra small
text-base lg:text-xs ( e.i. in the layout test... ).